+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changes by Version
 ==================
 
-0.17.5 (unreleased)
+0.17.5 (2015-10-12)
 -------------------
 
 - Set default checksum to ``CRC32C``.